ontrack-extension-indicators/src/test/java/net/nemerosa/ontrack/extension/indicators/values/BooleanIndicatorValueTypeTest.kt
package net.nemerosa.ontrack.extension.indicators.values

import com.fasterxml.jackson.databind.JsonNode
import com.fasterxml.jackson.databind.node.BooleanNode
import com.fasterxml.jackson.databind.node.NullNode
import net.nemerosa.ontrack.extension.indicators.IndicatorsExtensionFeature
import net.nemerosa.ontrack.extension.indicators.model.IndicatorCompliance
import net.nemerosa.ontrack.json.asJson
import net.nemerosa.ontrack.json.parseAsJson
import net.nemerosa.ontrack.model.structure.NameDescription
import org.junit.Test
import kotlin.test.assertEquals
import kotlin.test.assertNotNull

class BooleanIndicatorValueTypeTest {

    private val type = BooleanIndicatorValueType(IndicatorsExtensionFeature())

    @Test
    fun status() {
        assertEquals(IndicatorCompliance.MEDIUM, type.status(BooleanIndicatorValueTypeConfig(false), false))
        assertEquals(IndicatorCompliance.HIGHEST, type.status(BooleanIndicatorValueTypeConfig(false), true))
        assertEquals(IndicatorCompliance.LOWEST, type.status(BooleanIndicatorValueTypeConfig(true), false))
        assertEquals(IndicatorCompliance.HIGHEST, type.status(BooleanIndicatorValueTypeConfig(true), true))
    }

    @Test
    fun form() {

        fun assertForm(required: Boolean, value: Boolean?, expectedValue: String?) {
            val form = type.form(
                    NameDescription.nd("my-name", "My description"),
                    BooleanIndicatorValueTypeConfig(required),
                    value
            )
            assertNotNull(form.getField("value")) {
                assertEquals("my-name", it.label)
                assertEquals("My description", it.help)
                assertEquals(expectedValue, it.value)
            }
        }

        assertForm(required = false, value = null, expectedValue = "")
        assertForm(required = false, value = false, expectedValue = "false")
        assertForm(required = false, value = true, expectedValue = "true")
        assertForm(required = true, value = null, expectedValue = "")
        assertForm(required = true, value = false, expectedValue = "false")
        assertForm(required = true, value = true, expectedValue = "true")

    }

    @Test
    fun toClientJson() {

        fun assertClientJson(required: Boolean, value: Boolean, expectedValue: String) {
            val json = type.toClientJson(
                    BooleanIndicatorValueTypeConfig(required),
                    value
            )
            val expectedJson = mapOf(
                    "value" to expectedValue
            ).asJson()
            assertEquals(expectedJson, json)
        }

        assertClientJson(required = false, value = false, expectedValue = "false")
        assertClientJson(required = false, value = true, expectedValue = "true")
        assertClientJson(required = true, value = false, expectedValue = "false")
        assertClientJson(required = true, value = true, expectedValue = "true")

    }

    @Test
    fun fromClientJson() {

        fun assertClientJson(input: JsonNode, expectedValue: Boolean?) {
            val value = type.fromClientJson(BooleanIndicatorValueTypeConfig(false), input)
            assertEquals(expectedValue, value)
        }

        assertClientJson(NullNode.instance, null)
        assertClientJson("true".parseAsJson(), null)
        assertClientJson("1".parseAsJson(), null)
        assertClientJson(""""text"""".parseAsJson(), null)
        assertClientJson(mapOf("test" to true).asJson(), null)

        assertClientJson(mapOf("value" to null).asJson(), null)
        assertClientJson(mapOf("value" to false).asJson(), false)
        assertClientJson(mapOf("value" to true).asJson(), true)

    }

    @Test
    fun fromStoredJson() {

        fun assertStoredJson(input: JsonNode, expectedValue: Boolean?) {
            val value = type.fromStoredJson(BooleanIndicatorValueTypeConfig(false), input)
            assertEquals(expectedValue, value)
        }

        assertStoredJson(NullNode.instance, null)
        assertStoredJson("1".parseAsJson(), null)
        assertStoredJson(""""text"""".parseAsJson(), null)
        assertStoredJson(mapOf("test" to true).asJson(), null)

        assertStoredJson("true".parseAsJson(), true)
        assertStoredJson("false".parseAsJson(), false)

    }

    @Test
    fun toStoredJson() {

        fun assertStoredJson(input: Boolean, expectedNode: BooleanNode) {
            val stored = type.toStoredJson(BooleanIndicatorValueTypeConfig(false), input)
            assertEquals(expectedNode, stored)
        }

        assertStoredJson(false, BooleanNode.FALSE)
        assertStoredJson(true, BooleanNode.TRUE)

    }

    @Test
    fun configForm() {

        fun assertForm(config: BooleanIndicatorValueTypeConfig?, expectedValue: Boolean) {
            val form = type.configForm(config)
            assertNotNull(form.getField("required")) {
                assertEquals("Required", it.label)
                assertEquals(expectedValue, it.value)
            }
        }

        assertForm(null, expectedValue = true)
        assertForm(BooleanIndicatorValueTypeConfig(true), expectedValue = true)
        assertForm(BooleanIndicatorValueTypeConfig(false), expectedValue = false)

    }

    @Test
    fun toConfigForm() {

        assertEquals(
                mapOf("required" to false).asJson(),
                type.toConfigForm(BooleanIndicatorValueTypeConfig(false))
        )

        assertEquals(
                mapOf("required" to true).asJson(),
                type.toConfigForm(BooleanIndicatorValueTypeConfig(true))
        )

    }

    @Test
    fun fromConfigForm() {

        fun assertConfigForm(input: JsonNode, expectedRequired: Boolean) {
            val config = type.fromConfigForm(input)
            assertEquals(expectedRequired, config.required)
        }

        assertConfigForm(NullNode.instance, expectedRequired = true)
        assertConfigForm("1".parseAsJson(), expectedRequired = true)
        assertConfigForm(""""text"""".parseAsJson(), expectedRequired = true)
        assertConfigForm("true".parseAsJson(), expectedRequired = true)
        assertConfigForm("false".parseAsJson(), expectedRequired = true)
        assertConfigForm(mapOf("required" to 1).asJson(), expectedRequired = true)
        assertConfigForm(mapOf("required" to true).asJson(), expectedRequired = true)
        assertConfigForm(mapOf("required" to false).asJson(), expectedRequired = false)

    }

    @Test
    fun toConfigClientJson() {

        assertEquals(
                mapOf("required" to false).asJson(),
                type.toConfigClientJson(BooleanIndicatorValueTypeConfig(false))
        )

        assertEquals(
                mapOf("required" to true).asJson(),
                type.toConfigClientJson(BooleanIndicatorValueTypeConfig(true))
        )

    }

    @Test
    fun toConfigStoredJson() {

        assertEquals(
                mapOf("required" to false).asJson(),
                type.toConfigStoredJson(BooleanIndicatorValueTypeConfig(false))
        )

        assertEquals(
                mapOf("required" to true).asJson(),
                type.toConfigStoredJson(BooleanIndicatorValueTypeConfig(true))
        )

    }

    @Test
    fun fromConfigStoredJson() {

        fun assertFromStore(input: JsonNode, expectedRequired: Boolean) {
            val config = type.fromConfigStoredJson(input)
            assertEquals(expectedRequired, config.required)
        }

        assertFromStore(NullNode.instance, expectedRequired = true)
        assertFromStore("1".parseAsJson(), expectedRequired = true)
        assertFromStore(""""text"""".parseAsJson(), expectedRequired = true)
        assertFromStore("true".parseAsJson(), expectedRequired = true)
        assertFromStore("false".parseAsJson(), expectedRequired = true)
        assertFromStore(mapOf("required" to 1).asJson(), expectedRequired = true)
        assertFromStore(mapOf("required" to true).asJson(), expectedRequired = true)
        assertFromStore(mapOf("required" to false).asJson(), expectedRequired = false)


    }

}